+++ lib/PaymentTipAndPay/PaymentTipAndPayContractor.dart
@@ -8,7 +8,7 @@ abstract class PaymentTipAndPayContarctor {
     String orderType,
     int tableId,
     List items,
-    int totalAmount,
+    double totalAmount,
     String latitude,
     String longitude,
     BuildContext context,
